backend/requirements.txt
blinker==1.9.0
catboost==1.2.7
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
Flask==3.1.0
flask-cors
fonttools==4.55.8
fsspec==2025.2.0
google-generativeai==0.8.3
graphviz==0.20.3
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
narwhals==1.25.1
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.5.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
xgboost==1.7.6
python-dotenv==1.0.1